refactor: [Phase2] MVC 폴더로 위치 변경
