+++ docs/REQUIREMENTS.md
@@ -107,7 +107,8 @@
 
 ## 🎯 기능 요구 사항
 
-[ ] 자동차에 이름을 부여할 수 있다. 전진하는 자동차를 출력할 때 자동차 이름을 같이 출력한다.
+[ ] 자동차에 이름을 부여할 수 있다. 
+[ ] 전진하는 자동차를 출력할 때 자동차 이름을 같이 출력한다.
 [ ] 자동차 이름은 쉼표(,)를 기준으로 구분하며 이름은 5자 이하만 가능하다.
 [ ] 자동차 경주는 5회로 고정하여 진행한다.
 [ ] 전진하는 조건은 0에서 9 사이에서 무작위 값을 구한 후 무작위 값이 4 이상일 경우이다.
